+++ aitools/storage/base.py
